+++ src/vault/interact.rs
@@ -1,68 +1,18 @@
 use std::{cell::RefCell, collections::HashMap, error::Error, rc::Rc};
 
-use aes_gcm::{aead::OsRng, Aes256Gcm, Key};
-use argon2::password_hash::SaltString;
+use aes_gcm::{Aes256Gcm, Key};
 use inquire::Confirm;
-use serde::{Deserialize, Serialize};
 
 use crate::{
-    action::Record,
     cli::CLICommands,
     command::{Commands, Instructions, Interaction},
     file::{BackupFile, ProjectFile, RecordFile, SchemaFile, VaultFile},
     output::Output,
     schema::Schema,
     secure::{Encrypted, SecureData},
-    vault::Vault,
 };
 
-#[derive(Debug, Serialize, Deserialize)]
-pub struct PasswordEncrypted<Data> {
-    data: Encrypted<Data>,
-    salt: String,
-}
-
-impl<Data> SecureData for PasswordEncrypted<Data> {
-    type Item = Data;
-    fn salt(&self) -> &str {
-        &self.salt
-    }
-    fn data(&self) -> &Encrypted<Self::Item> {
-        &self.data
-    }
-}
-
-pub type VaultEncrypted = PasswordEncrypted<Vault>;
-
-impl VaultEncrypted {
-    fn new(password: String) -> Result<Self, Box<dyn Error>> {
-        let salt = SaltString::generate(&mut OsRng).to_string();
-        let key = Self::get_key(&salt, password);
-        Encrypted::encrypt(&Vault::new(), key).map(|vault| Self { data: vault, salt })
-    }
-
-    fn update(&mut self, data: &Vault, key: Key<Aes256Gcm>) -> Result<(), Box<dyn Error>> {
-        let updated = Encrypted::encrypt(data, key)?;
-        self.data = updated;
-        Ok(())
-    }
-}
-
-pub type RecordEncrypted = PasswordEncrypted<Record>;
-
-impl RecordEncrypted {
-    fn new(password: String) -> Result<Self, Box<dyn Error>> {
-        let salt = SaltString::generate(&mut OsRng).to_string();
-        let key = Self::get_key(&salt, password);
-        Encrypted::encrypt(&Record::new(), key).map(|vault| Self { data: vault, salt })
-    }
-
-    fn update(&mut self, data: &Record, key: Key<Aes256Gcm>) -> Result<(), Box<dyn Error>> {
-        let updated = Encrypted::encrypt(data, key)?;
-        self.data = updated;
-        Ok(())
-    }
-}
+use super::encrypted::{RecordEncrypted, VaultEncrypted};
 
 pub struct VaultInterface {
     vault_file: VaultFile,
